pyproject.toml
[tool.poetry]
name = "spotify"
version = "0.1.0"
description = "API tests for Spotify"
authors = ["Aleksandr Zavialov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = ">=7.3.1"
requests = ">=2.31.0"
jsonschema = ">=4.17.3"
allure-pytest = ">=2.13.2"
curlify = ">=2.2.1"
selene = ">=2.0.0.rc2"
pydantic = "1.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
